Moved much of the location setting to the viewmodel, where I wanted it. However, this requires passing in the permission isGranted. I'm checking it multiple times, but a bad eng could try to do this without checking, so it's surrounded in a try/catch. This also will request the permission immediately after the user grants it now.
